fix alarm state int -> name mapping. add user id filter. UI changes for small screens.
