+++ frontend/src/common/AppBar.tsx
@@ -12,7 +12,11 @@ export const OSDAppBar = () => {
     return (
         <Box sx={{ flexGrow: 1 }}>
             <AppBar position="static">
-                <Toolbar>
+                <Toolbar variant="dense">
+                    <img
+                        src="https://www.openseizuredetector.org.uk/wp-content/uploads/2015/02/icon_48x48.png"
+                        height={36}
+                    />
                     <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
                         OpenSeizureDetector
                     </Typography>
