+++ frontend/src/common/MultiSelectDropdown.tsx
@@ -6,22 +6,26 @@ import FormControl from '@mui/material/FormControl';
 import ListItemText from '@mui/material/ListItemText';
 import Select, { SelectChangeEvent } from '@mui/material/Select';
 import Checkbox from '@mui/material/Checkbox';
+import { Typography } from '@mui/material';
 
 export const MultiSelectDropdown = (props: MultiSelectDropdownProps) => {
     return (
         <div>
-            <FormControl sx={{ m: 1, width: 300 }}>
+            <FormControl sx={{ m: 1, width: 200 }} size="small">
                 <InputLabel>{props.label}</InputLabel>
                 <Select
                     multiple
+                    size="small"
                     value={props.selected}
                     onChange={props.handleChange}
                     input={<OutlinedInput label={props.label} />}
-                    renderValue={(selected: string[]) => selected.join(', ')}>
+                    renderValue={(selected: string[]) => (
+                        <Typography sx={{ textAlign: 'start' }}>{selected.join(', ')}</Typography>
+                    )}>
                     {props.options.map((option) => (
-                        <MenuItem key={option} value={option}>
-                            <Checkbox checked={props.selected.indexOf(option) > -1} />
-                            <ListItemText primary={option} />
+                        <MenuItem key={option} value={option} dense>
+                            <Checkbox size="small" checked={props.selected.indexOf(option) > -1} />
+                            <ListItemText primary={option} sx={{ fontSize: 8 }} />
                         </MenuItem>
                     ))}
                 </Select>
