+++ frontend/src/router/Router.tsx
@@ -2,7 +2,7 @@ import { useContext } from 'react';
 import { AuthStateContext } from '../context/AuthStateContext';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { Login } from '../pages/Login';
-import { Home } from '../pages/Home';
+import { DataTablePage } from '../pages/dataTablePage/DataTablePage';
 import { Register } from '../pages/Register';
 
 export const Router = () => {
@@ -20,6 +20,6 @@ const PublicRoute = () => (
 
 const ProtectedRoute = () => (
     <Routes>
-        <Route path="/" element={<Home />} />
+        <Route path="/" element={<DataTablePage />} />
     </Routes>
 );
